Change $(PROJECT_DIR) to $(SRCROOT).
Using SwiftBase58 as a subproject breaks with the project_dir environment variable
 because when compiling the super project all the Carthage checkouts are kept in
the super/root project directory. Say I'm building swift-ipfs-api and have SwiftBase58
as a subproject. When SwiftBase58 looks in its Carthage/Checkouts/SwiftGMP/GMPLib dir
for a lib it *should* look for it in the swift-ipfs-api/Carthage/Checkouts/SwiftGMP/GMPLib
instead. This is because Carthage doesn't put subprojects' checkouts in their Carthage dir
but collects them all at the root of the super/root project.
The (not much) hope is that by replacing the SwiftBase58 lib path with $srcroot it will
look at the root of the super project instead.